% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{bam.read}
\alias{bam.read}
\title{bam.read}
\usage{
bam.read(bam, out.file = NULL, region = NULL, fields = c("seq"),
  tags = c("MD", "BX"), nlines = NULL, estimate.lines = TRUE,
  skip = NULL, chunksize = 1e+05, mc.cores = 1, min.mapq = 0,
  samtools.flag = "", verbose = TRUE)
}
\arguments{
\item{bam}{bam file}

\item{out.file}{file to dump csv to, if null returns data.table}

\item{tags}{which tags to extract from bam file}

\item{nlines}{number of lines in bam file if known, otherwise will be estimated}

\item{estimate.lines}{flag whether to count lines or just estimate using bam.size if nlines is not provided (default = TRUE)}

\item{skip}{how many lines to skip}

\item{mc.cores}{how many threads to use for parsing}

\item{min.mapq}{integer min mapping quality}

\item{samtools.flag}{additoinal samtools flag to use}

\item{verbose}{flag}
}
\description{
Loads entire bam or slice into data.table or dumps to csv via samtools call (instead
of Rsamtools like read.bam)
}
\author{
Marcin Imielinski
}
