% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{bam2bw}
\alias{bam2bw}
\title{bam2bw}
\usage{
bam2bw(files.bam, files.bw = NULL, fork = 0)
}
\description{
Runs BEDTools genomeCoverageBed and then UCSC tool BedGraphToWigWig
on a set of bam files to output per base coverage in bigwig format.

If files.bw is not provided then will just replace .bam with .bedtools.coverage.bw in files.wig name)

NOTE: assumes that BedGraphToBigWig and genomeCoverageBed already exist on file system
and skidb_env('UCSC.CHROM.SIZES.PATH') should be valid.   Also assumes bam is sorted (ie samtools sort bam)

reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}

