% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{blastp}
\alias{blastp}
\title{blastp}
\usage{
blastp(seq, db = skidb_env("UNIPROT.FA"))
}
\description{
Run blastp (via command line) on (single) sequence against uniprot DB (default) or arbitrary DB
}
\author{
Marcin Imielinski
}
