% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{bw2rle}
\alias{bw2rle}
\title{bw2rle}
\usage{
bw2rle(files.bw, regions = NULL, files.rle = NULL, precision = 2,
  fork = 0, return.objects = TRUE, dump.files = TRUE, verbose = FALSE)
}
\description{
convert each big wig (bw) file in files.bw into an RDS file storing a SimpleRleList
across list of regions (GRanges or seg data frame)

if regions is NULL will process all data stored in the bw file (warning can be large)

These files can be loaded with readRDS
}
\author{
Marcin Imielinski
}
