% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{dreme}
\alias{dreme}
\title{dreme}
\usage{
dreme(case, control = NULL, basedir = "./", tilim = 1000, mink = 3,
  maxk = 8, e.thresh = 0.05, m.thresh = NULL, ngen = 100,
  MEMEPATH = "/nfs/sw/meme/meme-4.11.2/bin/dreme")
}
\description{
run dreme on biostrings case control input with given parameters
}
