% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{fh_containers}
\alias{fh_containers}
\alias{fh_getcontainers}
\title{fh_containers}
\usage{
fh_getcontainers(ids, id.type = "Sample", container.type = "Individual",
  wkspace = Sys.getenv("FIREHOSE_WORKSPACE"), verbose = FALSE)
}
\description{
(deprecated pre-fiss function)

call to getcontainers firehose API function for a set of entity ids (eg Sample Id)

id entity type should match the entity ids (ie if sample ids should be "Sample", if individual ids should be "Individual")
}
\author{
Marcin Imielinski
}

