% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{fh_import}
\alias{fh_import}
\title{fh_import}
\usage{
fh_import(name, inds, wkspace = Sys.getenv("FIREHOSE_WORKSPACE"))
}
\description{
Import individual set with name iset name "name" and individuals "inds" into firehose

NOTE: requires that a file ~/.fh.config exists in home directory with a single line: "-u username:password"
(where username = fh username, password = fh password) --> this file should be "chmod 700" to maintain (some) pw security
}
\author{
Marcin Imielinski
}

