% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{fh_start}
\alias{fh_start}
\title{fh_start}
\usage{
fh_start(pipeline, isetname = DEFAULT.ISET,
  wkspace = Sys.getenv("FIREHOSE_WORKSPACE"))
}
\description{
Pushes button of firehose pipeline "pipeline" in individual set "isetnmae" in workspace "wkspace"

NOTE: requires that a file ~/.fh.config exists in home directory with a single line: "-u username:password"
(where username = fh username, password = fh password) --> this file should be "chmod 700" to maintain (some) pw security
}
\author{
Marcin Imielinski
}
