% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_delete}
\alias{fiss_delete}
\title{fiss_delete}
\usage{
fiss_delete(entities = NULL, type = "Individual",
  space = Sys.getenv("FIREHOSE_WORKSPACE"), wkspace = space, char = F,
  domain = "cga")
}
\description{
Delete entities via fiss
}
\author{
Marcin Imielinski
}

