% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_get}
\alias{fiss_get}
\title{fiss_get}
\usage{
fiss_get(set = NULL, space = Sys.getenv("FIREHOSE_WORKSPACE"),
  wkspace = space, type = "Pair", sample_type_col = "sample_type",
  fuse = TRUE, verbose = F, get.containers = T, mc.cores = 1,
  domain = "cga")
}
\description{
replacement for get fh

gets all the individual / pair and sample info for an individual / pair set
}
\author{
Marcin Imielinski
}

