% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_getcontainers}
\alias{fiss_getcontainers}
\title{fiss_getcontainers}
\usage{
fiss_getcontainers(id, type = "sample",
  wkspace = Sys.getenv("FIREHOSE_WORKSPACE"), verbose = F, mc.cores = 1,
  domain = "cga")
}
\description{
low level function to get containers
}
\author{
Marcin Imielinski
}

