% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_import}
\alias{fiss_import}
\title{fiss_import}
\usage{
fiss_import(dat, type = "pair", space = Sys.getenv("FIREHOSE_WORKSPACE"),
  wkspace = space, domain = "cga")
}
\description{
import individual, sample, or pair annotation to firehose workspace via fiss, writing to temp tables
(NOTE: importing a sample annotation = annotating a sample, or setting an annotation for a sample)

Value of dat depends on value of type

if type = 'sample' or 'pair' then dat is data.frame whose rows are named by the individual or pair_id, depending the value of type, i.e.
"pair_id" if type = 'Pair', 'sample_id' if type = 'Sample', and the remaining columns are annotations to add
NOTE: annotations already have to be declared from within firehose, not currently possibly to do from FISS

if type = 'iset', 'pset', or 'sset' then dat is a two column data.frame with first column having set name and second column having the
individuals / pairs / samples belonging to it.

NOTE: if type = 'wkspace' then dat is just a named vector specifying workspace annotations
}
\author{
Marcin Imielinski
}
