% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_list}
\alias{fiss_list}
\title{fiss_list}
\usage{
fiss_list(type = "pset", set = NULL,
  space = Sys.getenv("FIREHOSE_WORKSPACE"), wkspace = space,
  verbose = FALSE, domain = "cga")
}
\description{
List all entities of a given type in a given workspace.
}
\author{
Marcin Imielinski
}

