% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_new_set}
\alias{fiss_new_set}
\title{fiss_new_set

creates individual, pair, sample, pair set, iset entities for an input sample set in given wkspace

Input table samples must have fields
$sample_id
$individual_id

and can have optional fields
$pair_id, $case (logical), each pair_id must occur with one $case = TRUE, and one $case = FALSE
$pset_id
$iset_id

as well as other sample level annotations associated with the sample (eg bam_file_wgs)}
\usage{
fiss_new_set(samples, space, verbose = T)
}
\description{
List all entities of a given type in a given workspace.
}
\author{
Marcin Imielinski
}

