% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_status}
\alias{fh_status}
\alias{fiss_status}
\title{fiss_status}
\usage{
fh_status(workflow = "Capture_Workflow", isetname = DEFAULT.ISET,
  wkspace = Sys.getenv("FIREHOSE_WORKSPACE"))
}
\description{
Gets workflow status for iset in workspace.

NOTE: requires that a file ~/.fh.config exists in home directory with a single line: "-u username:password"
(where username = fh username, password = fh password) --> this file should be "chmod 700" to maintain (some) pw security
List all entities of a given type in a given workspace.
}
\author{
Marcin Imielinski
}

