% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{fiss_task}
\alias{fiss_task}
\title{fiss_task}
\usage{
fiss_task(name = NULL, entity = NULL, type = "pset", start = F,
  stop = F, export = FALSE, verbose = F, domain = "cga",
  space = Sys.getenv("FIREHOSE_WORKSPACE"), wkspace = space)
}
\description{
Run a task within firehose via fiss
}
\author{
Marcin Imielinski
}

