% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{gaf2gr}
\alias{gaf2gr}
\title{gaf2gr}
\usage{
gaf2gr(gaf, strip.chr = T)
}
\description{
converts one or more gaf data frame rows to GRanges list (works only for gaf rows with composite type genome) using "CompositeCoordinates"
named by FeatureID
}
\author{
Marcin Imielinski
}

