% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{get_alignment_summaries}
\alias{get_alignment_summaries}
\alias{process_uniprot_ft}
\title{get_alignment_summaries}
\usage{
get_alignment_summaries(iset, wgs = T, short = FALSE)

process_uniprot_ft(out.file = skidb_env("UNIPROT.FT"),
  uniprot.dat = skidb_env("UNIPROT.DAT"))
}
\description{
get_alignment_summaries

Using firehose table grabs the picard alignment summaries and pulls out samplewise statistics for tumor and normal
including pf_reads, pf_aligned_bases etc.

Also computes ballpark X coverage for wgs and wes using back of the envelope sizes for each (34MB, 3GB)

Returns table with Tumor_ and Normal_ columns added at the end
call to getcontainers firehose API function for a set of entity ids (eg Sample Id)

process_uniprot_ft

Takes uniprot data file and outputs a "seg" file style data frame / txt of protein features file with columns $ID, $begin, $end,  $feature.name, $feature.description
}
\author{
Marcin Imielinski

= functioun Marcin Imielinski
}
