% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{get_fh}
\alias{get_fh}
\title{get_fh}
\usage{
get_fh(isetname, wkspace = "An_LUAD", type = "Individual", merged = TRUE,
  hack.sample.mapping = T, verbose = FALSE)
}
\description{
Deprecated (pre-fiss) function of accessing firehose

Retrieves sample or individual table from firehose for individual set isetname and workspace wkspace.
NOTE: isetname can be "All", in which case all individuals in that workspace are retrieved.

If isetname, wkspace don't exist will yield a two row table that contains errors.

NOTE: requires that a file ~/.fh.config exists in home directory with a single line: "-u username:password"
(where username = fh username, password = fh password) --> this file should be "chmod 700" to maintain (some) pw security
}
\author{
Marcin Imielinski
}
