% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{get_ucsc_bands}
\alias{get_ucsc_bands}
\title{get_ucsc_bands}
\usage{
get_ucsc_bands(hg19 = T, arms = F, add.chr = F, gr = T)
}
\description{
returns segments corresponding to coordinates of ucsc hg18 / hg19 chrom bands
}
\author{
Marcin Imielinski
}

