% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{get_ucsc_chroms}
\alias{get_ucsc_chroms}
\title{get_ucsc_chroms}
\usage{
get_ucsc_chroms(hg19 = T, clean = T, numeric = F, add.chr = F)
}
\description{
returns segments corresponding to coordinates of ucsc hg18 / hg19 chromosomesfour

"clean" will return without all of the unmapped segments (e.g Un_gl000221)
reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}
