% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{mdb}
\alias{mdb}
\title{mdb}
\usage{
mdb(fn, where = NULL, name = "data_table", all = FALSE, con = FALSE)
}
\arguments{
\item{fn}{filename or DBI connection to sqllite}

\item{where}{expression on which to filter}
}
\description{
Quickly queries MonetDB file or connection using expressions on columns to select, returning
data.table
}
\author{
Marcin Imielinski
}
