% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{meme}
\alias{meme}
\title{meme}
\usage{
meme(sequences, basedir = "./", tilim = 18000, maxsize = 60000,
  nmotifs = 8, minw = 3, maxw = 50, revcomp = TRUE,
  MEMEPATH = "/nfs/sw/meme/meme-4.11.2/bin/meme")
}
\description{
run meme on biostring input
}
\author{
Marcin Imielinski
}
