% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{moglow}
\alias{moglow}
\title{moglow}
\usage{
moglow(motif, case, thresh = "80\%", fn = "~/public_html/dump.html")
}
\arguments{
\item{motif}{length 1 character vector representing motif to}

\item{case}{character vector sequences to plot}
}
\description{
Plots motif instances on sequences as HTML
}
\author{
Marcin Imielinski
}
