% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{muscle}
\alias{muscle}
\title{muscle}
\usage{
muscle(seq, sub.mat = NULL, gap.open = -10, gap.extend = -4,
  html.out = NULL)
}
\description{
runs muscle program on XStringSet (e.g. AAStringSet, DNAStringSet) object "seq" and
return output as XStringSet

reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}
