% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{process_gaf}
\alias{process_gaf}
\title{process_gaf}
\usage{
process_gaf(gaf.file = skidb_env("GAF.SOURCE"), verbose = T)
}
\description{
processes gaf from source (ie obtained from TCGA) into skidb_env('GAF.DIR')
}
\author{
Marcin Imielinski
}

