% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{process_pfam}
\alias{process_pfam}
\title{process_pfam}
\usage{
process_pfam(pfam.file = skidb_env("PFAM.HUMAN"),
  out.file = skidb_env("PFAM.HUMAN.RDS"), lim = Inf)
}
\description{
Given pfam msa file (stockholm format)
output data as list of character matrices with rownames corresponding to sequence data
and list names corresponding to pfam domain IDs.
}
\author{
Marcin Imielinski
}

