% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{process_pfam_metaa}
\alias{process_pfam_meta}
\alias{process_pfam_metaa}
\title{process_pfam_meta}
\usage{
process_pfam_meta(pfam.fa.file = skidb_env("PFAM.FA"),
  outfile = skidb_env("PFAM.HUMAN.META.RDS"), verbose = T)
}
\description{
processes pfam fasta from pfam fasta and main pfam file
}
\author{
Marcin Imielinski
}

