% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{process_repeatmasker}
\alias{process_repeatmasker}
\title{process_repeatmasker}
\usage{
process_repeatmasker(infile = skidb_env("REPEATMASKER.HG19"),
  outfile = skidb_env("REPEATMASKER.HG19.GR"), chrsub = T)
}
\description{
process repeat masker file into GRanges
}
\author{
Marcin Imielinski
}
