% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{process_uniprot_acmap}
\alias{process_uniprot_acmap}
\title{process_uniprot_acmap}
\usage{
process_uniprot_acmap(out.file = skidb_env("UNIPROT.ACMAP"),
  uniprot.dat = skidb_env("UNIPROT.DAT"))
}
\description{
Takes uniprot data file and outputs a two column table mapping id's to accession numbers
}
\author{
Marcin Imielinski
}
