% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{process_uniprot_rs}
\alias{process_uniprot_rs}
\title{process_uniprot_rs}
\usage{
process_uniprot_rs(out.file = skidb_env("UNIPROT.RS"),
  uniprot.dat = skidb_env("UNIPROT.DAT"))
}
\description{
Takes uniprot data file and outputs a two column table mapping id's to refseq numbers
}
\author{
Marcin Imielinski
}
