% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read.xml}
\alias{read.xml}
\title{read.xml}
\usage{
read.xml(files, pattern = NULL, as.data.table = TRUE, na.sort = TRUE,
  mc.cores = 1)
}
\description{
reads a bunch of xml in files into data.table, flattens, and
and concatenates the files via rbindlist fill
}
