% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_10x}
\alias{read_10x}
\title{tally 10x fragments across windows}
\usage{
read_10x(bam = NULL, wins, reads = NULL, cocov = TRUE,
  readcount = FALSE, verbose = TRUE)
}
\arguments{
\item{bam}{bam file}

\item{wins}{GRanges of windows to query and tally}

\item{cocov}{logical flag whether to return cocov length(win) x length(win) matrix or matrix of fragments x windows}
}
\description{
pull reads from 10x .bam across windows and return matrix barcodes x windows or window x window co-coverage matrix
}
