% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_ensembl}
\alias{read_ensembl}
\title{read_ensembl}
\usage{
read_ensembl(transcript = F, flat = F, grl = F, types = "")
}
\description{
reads ensembl transcripts into genomic ranges
}
\author{
Marcin Imielinski
}

