% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_gaf}
\alias{read_gaf}
\title{read_gaf}
\usage{
read_gaf(file = "gene", gchain = F)

read_gaf(file = "gene", gchain = F)
}
\description{
reads any one of the gaf rds files stored in skidb_env('GAF.DIR') by process_gaf

returns GRanges of genome coordinates of gaf features or gChain mapping feature coordinates to composite coordinates
}
\author{
Marcin Imielinski

Marcin Imielinski
}
