% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_gencode}
\alias{read_gencode}
\title{read_gencode}
\usage{
read_gencode(type = NULL, by = NULL,
  fn.rds = skidb_env("GENCODE.FILE.HG19.GR"), fn = GENCODE.FILE.HG19)
}
\description{
reads GENCODE file dump from .rds of GRanges.

If that file doesn't exist
}
\author{
Marcin Imielinski
}

