% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_kg}
\alias{read_kg}
\title{read_kg}
\usage{
read_kg(hg19 = T, chr = F, gr = F, grl = F)
}
\description{
reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}

