% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_lumpy}
\alias{read_lumpy}
\title{read_lumpy}
\usage{
read_lumpy(file, sl = hg_seqlengths(), samples = c("tumor", "normal"))
}
\description{
reads LUMPY sv bedpe format, returning rearrangement junctions as GRangesList
with strands pointing AWAY from junction
}
\author{
Marcin Imielinski
}

