% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_maf}
\alias{read_maf}
\title{read_maf}
\usage{
read_maf(fn, nskip = NULL, cols = NULL, dt = FALSE, add.path = FALSE,
  verbose = FALSE)
}
\description{
read maf taking care of headers
}
\author{
Marcin Imielinski
}
