% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_msigdb}
\alias{read_msigdb}
\title{read_msigdb}
\usage{
read_msigdb(path = skidb_env("MSIGDB_PATH"))
}
\description{
reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}

