% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_peaks}
\alias{read_peaks}
\title{read_peaks}
\usage{
read_peaks(fn, chr.sub = TRUE)
}
\description{
reads UCSC / ENCODE / Roadmap broad and narrow peak data into granges object
}
\author{
Marcin Imielinski
}
