% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_refgene}
\alias{read_refGene}
\alias{read_refgene}
\title{read_refgene}
\usage{
read_refGene(hg19 = TRUE, exons = F, cds = F, granges = F, grl = F,
  chr.numeric = F, chr = F, uniprot = T, ccds = T, force.make = F)
}
\description{
if exons = T, returns an expanded data frame with 1 row per exon
if uniprot = T, maps to uniprot annotations, and limits only to those refseq genes that map to at least
one uniprot record
}
\author{
Marcin Imielinski
}

