% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_snpdata}
\alias{read_snpdata}
\title{read_snpdata}
\usage{
read_snpdata(file, markermap = skidb_env("SNP.MARKER.MAP"), td = FALSE)
}
\description{
reads snp data given array definition file in either rds data table format
or txt format

returns GRanges or list with $gr and $td for GRanges and length(2) trackData of total and allelic copy number

process rsem files from UNC
}
\author{
Marcin Imielinski
}

