% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_stockholm}
\alias{read_stockholm}
\title{read_stockholm}
\usage{
read_stockholm(pfam.file, lim = Inf)
}
\description{
Reads Pfam stockholm format into a list of matrices named by each record accession
Each row of each matrix comprises a sequence alignment.
}
\author{
Marcin Imielinski
}
