% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_ucsc_files_table}
\alias{read_ucsc_files_table}
\title{read_ucsc_files_table}
\usage{
read_ucsc_files_table(url)
}
\description{
parses contents of url pointing to files.txt file on UCSC server
corresponding to track metadata and the corresponding files (which are
assumed to lie in the same directory

e.g. "http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeUwRepliSeq/files.txt"#'
}
\author{
Marcin Imielinski
}

