% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_uniprot_fa}
\alias{read_uniprot_fa}
\title{read_uniprot_fa}
\usage{
read_uniprot_fa(trim.names = T)
}
\description{
reads uniprot fasta file into BioStrings BStringSet format
}
\author{
Marcin Imielinski
}
