% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_uniprot_seqinfo}
\alias{read_uniprot_seqinfo}
\alias{uniprot_seqinfo}
\title{read_uniprot_seqinfo}
\usage{
uniprot_seqinfo()
}
\description{
extracts seqinfo from uniprot fa file
}
\author{
Marcin Imielinski
}

