% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{read_vcf}
\alias{read_vcf}
\title{read_vcf}
\usage{
read_vcf(fn, gr = NULL, hg = "hg19", geno = NULL, swap.header = NULL,
  verbose = FALSE, add.path = FALSE, tmp.dir = "~/temp/.tmpvcf", ...)
}
\description{
wrapper around variatnAnnotation reads VCF into granges or data.table format
}
\author{
Marcin Imielinski
}

