% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{rg2exons}
\alias{rg2exons}
\title{rg2exons

extracts a df list of exons from one or more transcripts of the rg data frame
reorient = T, numbers exons relative to transcript start.}
\usage{
rg2exons(rg, reorient = FALSE, cds = FALSE, gr = FALSE)
}
\description{
reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}

