% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{run_prego}
\alias{run_prego}
\alias{read_prego}
\title{run_prego}
\usage{
read_prego(fn)
}
\description{
writes RA and CN output to directory, so it can be run from command line

intervals = GRanges tiling of genome with $tum.cov and $norm.cov fields (obtain using bam.cov.gr)
ra = output of ra_breaks, GRangesList of signed breakpoint pairs pointing in the direction of the joined segment
}
\author{
Marcin Imielinski
}
