% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{skidb_env}
\alias{skidb_env}
\title{skidb_env}
\usage{
skidb_env(query = NULL)
}
\description{
Get environment variables of all flat file paths associated with skidb, or specific file path.  T

This can be used to
populate the file pointers with the appropriate files or rewire the paths using environment variables
having the same name as the skidb_env environment variable.  skidb_env is used by other skidb functions
to grab files and paths to software.


gets all the individual / pair and sample info for an individual / pair set
}
\author{
Marcin Imielinski
}

