% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{wig2bw}
\alias{wig2bw}
\title{wig2bw}
\usage{
wig2bw(files.wig, files.bw = NULL,
  chrom.path = skidb_env("UCSC.CHROM.SIZES.PATH"), fork = 0)
}
\description{
Runs UCSC command tool wig2big on a set of wig files (files.wig) and outputs to the corresponding bigwig files (files.bw)
wig files can be zipped as well

If files.bw is not provided then will just replace wig with big wig in files.wig name)
}
\author{
Marcin Imielinski
}
