% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{wig2covered}
\alias{wig2covered}
\title{wig2covered}
\usage{
wig2covered(fn, out.fn = paste(fn, ".covered.txt", sep = ""), verbose = T,
  gz = T)
}
\description{
takes wigfile of coverage with path fn (which is either a text, gz, or bz2 file) and
outputs all "covered" positions (chr pos) to out.fn as space separated file.

NOTE: assumes all positions without 0 value are covered
reads knownGene and refGene and crosses the two (left join) to yield an uber table
}
\author{
Marcin Imielinski
}

