% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skidb.R
\name{wig2rlea}
\alias{wig2rle}
\alias{wig2rlea}
\title{wig2rle}
\usage{
wig2rle(files.wig, files.rle, fork = 0, return.objects = TRUE,
  dump.files = TRUE, verbose = FALSE)
}
\description{
convert each big wig (bw) file in files.bw into an RDS file storing a SimpleRleList
across list of regions (GRanges or seg data frame)

These files can be loaded with readRDS
}
\author{
Marcin Imielinski
}

