% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{write.mdb}
\alias{write.mdb}
\title{write.mdb}
\usage{
write.mdb(
  dt,
  fn = tempdir(),
  name = "data_table",
  max.varchar = 256,
  index_on = NULL,
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dt}{(keyed) data.table}

\item{fn}{path to write to}
}
\value{
DBI handle to MDB
! @param index_on list of character vectors specifying additional keys (in addition to key(dt))
}
\description{
Writes data.table (or anything coercible to one) to MonetDB.  If data.table has key will create an index.
Additional indices may be specified.  If MonetDB file already exists then will try to append to it (assuming same schema).
Stored table will have altered column names (changed to lowercase, all non word characters replaced by underscores, and
a col_ prefix added) to maximize compatibility with SQL.

The written database can be later accessed using mdb(fn, expression) or mdb(con, expression) where con is the output of
this function (write.mdb) or con = mdb(fn).
}
\author{
Marcin Imielinski
}
