% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skidb.R
\name{meme}
\alias{meme}
\title{meme}
\usage{
meme(
  sequences,
  basedir = "./",
  alphabet = NULL,
  tilim = 18000,
  maxsize = 60000,
  nmotifs = 8,
  minw = 3,
  anr = FALSE,
  maxw = 50,
  revcomp = TRUE,
  mc.cores = 1,
  MEMEPATH = "/gpfs/commons/home/mimielinski/software/meme/bin/meme"
)
}
\description{
run meme on biostring input
}
\author{
Marcin Imielinski
}
